approve setting finish

+++ src/components/ApproveSetting/ApproveSetting.tsx
@@ -2,10 +2,11 @@ import { Component, PropSync, Prop, Watch, Emit } from "vue-property-decorator";
 import * as tsx from "vue-tsx-support";
 import './style/index.less';
 import customModel from '@/components/CustomModel/approveSettingModel';
-import { firstOptions, middleOptions, lastOptions } from './modelFormOptions';
+
 import { approveSettingItem } from '@/utils/interface'
-import { List } from 'ant-design-vue';
+import { selectList } from '@/api/role';
 import { setValue, deepCopy } from '@/utils/util';
+import { formPageTypeEnum as pageTypeEnum } from '@/utils/enum';
 
 // 定义节点类型，分别为发起人节点、审批人节点、最终审批人节点。不同节点有不同的模板
 enum nodeType {
@@ -24,11 +25,22 @@ enum theme {
 })
 export default class extends tsx.Component<Vue> {
 
+  @Prop({
+    type: Boolean,
+    default: false
+  }) public isDelete!: Boolean; // 是否是删除审批流设置
+
   // 从父组件接收到的审批流列表
   @Prop({
     type: Array,
     default: () => []
   }) public flowList!: Array<approveSettingItem>;
+
+  @Watch('flowList', {deep: true, immediate: true})
+  public watchFlowList(newValue: any) {
+    console.log("flowList发生变化", newValue);
+    this.list = newValue;
+  }
 
   // 组件内的审批流列表 为了解耦
   public list: Array<approveSettingItem> = [];
@@ -39,8 +51,16 @@ export default class extends tsx.Component<Vue> {
   public theme: theme = theme.blue;
   public activeIndex: number = 0; // 当前正在设置的审批项的index 
 
+  public roleList: any = [];  // 角色列表
+
+  public firstOptions! : any;
+  public middleOptions! : any;
+  public lastOptions! : any;
+  public secondOptions! : any;
+
   // 点击设置审批人
   public setPerson(index: number, list: Array<any>) {
+    var  { firstOptions, middleOptions, lastOptions, secondOptions }  = this;
     // 第一节点
     if (index === 0) {
       this.formOptions = setValue(firstOptions, list[index]);
@@ -49,6 +69,10 @@ export default class extends tsx.Component<Vue> {
       // 最后一个节点
       this.theme = theme.pink;
       this.formOptions = setValue(lastOptions, list[index]);
+    } else if(index == 1) {
+      // 第二个节点和中间节点相比没有 【审批未到当前节点时，该审批人能否查看本表单】 这一选项
+      this.theme = theme.orange;
+      this.formOptions = setValue(secondOptions, list[index]);
     } else {
       // 中间节点
       this.theme = theme.orange;
@@ -69,15 +93,39 @@ export default class extends tsx.Component<Vue> {
 
   // 删除某节点
   public deleteNode(index: number) {
-    this.list.splice(index,1);
+    this.list.splice(index, 1);
   }
 
   get id() {
     return this.$route.params.id;
   }
 
+  public get pageType() {
+    const { path } = this.$route;
+    if (this.id == 'create' || isNaN(Number(this.id))) {
+      return pageTypeEnum.create;
+    } else if (path.indexOf('detail') != -1) {
+      return pageTypeEnum.detail;
+    } else {
+      return pageTypeEnum.update;
+    }
+  }
+
   public handleSync(key: any, value: any) {
     this.$set(this, key, value)
+  }
+
+  // 获取角色名
+  public getRoleName(roles: Array<number>) {
+    if (Array.isArray(roles)) {
+      var roleNames: string[] = [];
+      this.roleList.map((item: any) => {
+        if (roles.indexOf(Number(item.value)) != -1) {
+          roleNames.push(item.label)
+        }
+      })
+      return roleNames.join('、');
+    }
   }
 
   // 点击设置节点表单提交按钮，更新数据项
@@ -89,10 +137,10 @@ export default class extends tsx.Component<Vue> {
   // 获取表单权限的命名
   public getFormRoleName(value: Array<number> | number, index: number) {
     var result = [];
-    if(index == 0 && Array.isArray(value)) {
-      for(let i in firstOptions[2].selectOptions) {
-        if(value.indexOf(firstOptions[2].selectOptions[i].value) != -1) {
-          result.push(firstOptions[2].selectOptions[i].label);
+    if (index == 0 && Array.isArray(value)) {
+      for (let i in this.firstOptions[2].selectOptions) {
+        if (value.indexOf(this.firstOptions[2].selectOptions[i].value) != -1) {
+          result.push(this.firstOptions[2].selectOptions[i].label);
         }
       }
     } else {
@@ -103,10 +151,10 @@ export default class extends tsx.Component<Vue> {
 
   // 获取审批权限的命名
   public getApproveRoleName(value: number[]) {
-    var selectOptions = middleOptions[middleOptions.length - 1].selectOptions;
+    var selectOptions = this.middleOptions[this.middleOptions.length - 1].selectOptions;
     var result = [];
-    for(let i in selectOptions) {
-      if(value.indexOf(selectOptions[i].value) != -1) {
+    for (let i in selectOptions) {
+      if (value.indexOf(selectOptions[i].value) != -1) {
         result.push(selectOptions[i].label);
       }
     }
@@ -118,6 +166,18 @@ export default class extends tsx.Component<Vue> {
     if (this.id === 'create') {
       this.list = [{}, {}, {}, {}];
     }
+    selectList().then(res => {
+      this.roleList = res;
+    })
+    var options = require('./modelFormOptions');
+    if(this.isDelete) {
+      var options = require('./delModelFormOptions');
+    }
+    var  { firstOptions, middleOptions, lastOptions, secondOptions }  = options;
+    this.firstOptions = firstOptions;
+    this.middleOptions = middleOptions;
+    this.lastOptions = lastOptions;
+    this.secondOptions = secondOptions;
   }
 
   protected render() {
@@ -127,17 +187,17 @@ export default class extends tsx.Component<Vue> {
         {
           list.map((item, index) => {
             let icon = [];
-            if((Array.isArray(item.roles) && item.roles.length > 0)) {
-              icon.push(<a-icon type="form" class="editIcon" onClick={this.setPerson.bind(this, index, list)}/>)
+            if ((Array.isArray(item.roles) && item.roles.length > 0) && this.pageType !== pageTypeEnum.detail) {
+              icon.push(<a-icon type="form" class="editIcon" onClick={this.setPerson.bind(this, index, list)} />)
             }
-            if (index !== 0) {
-              icon.push( <a-popconfirm
+            if (index !== 0 && this.pageType !== pageTypeEnum.detail) {
+              icon.push(<a-popconfirm
                 title="确定要删除这个节点吗?"
                 ok-text="确定"
                 cancel-text="不了"
                 on-confirm={this.deleteNode.bind(this, index)}
               ><a-icon type="delete" class="delete" /></a-popconfirm>)
-            } 
+            }
             return (
               <div class={['box-single', index === list.length - 1 && index !== 0 ? 'last' : '']}>
                 <div class="single">
@@ -154,11 +214,12 @@ export default class extends tsx.Component<Vue> {
                         (Array.isArray(item.roles) && item.roles.length > 0) ? (
                           <div class="roleList">
                             <img src={require('@/assets/img/person.png')} />
-                            <span>{ item.roles?.join(' 、 ') }</span>
+                            <span>{this.getRoleName(item.roles)}</span>
                           </div>
                         ) : (
-                          <div class="setPerson" onClick={this.setPerson.bind(this, index, list)}><a-icon type="setting" class="icon" /> 设置{index === 0 ? '发起人' : index === list.length - 1 ? '最终审批人' : '审批人'} </div>
-                        )
+                          this.pageType !== pageTypeEnum.detail &&
+                            <div class="setPerson" onClick={this.setPerson.bind(this, index, list)}><a-icon type="setting" class="icon" /> 设置{index === 0 ? '发起人' : index === list.length - 1 ? '最终审批人' : '审批人'} </div>
+                          )
                       }
                     </div>
                   </div>
@@ -170,8 +231,8 @@ export default class extends tsx.Component<Vue> {
                           temp.push(<div class="text">发起人<span>{item.ifChoose == 1 ? '能' : '不能'}</span>自选下一节点审批用户</div>)
                         } if (typeof item.approveType != 'undefined') {
                           temp.push(<div class="text">审批类型：<span>{item.approveType == 1 ? '一名审批用户通过/退回/否决即可' : '所有审批用户都需通过'}</span></div>)
-                        } if(typeof item.readBefore != 'undefined') {
-                          temp.push(<div class="text">审批未到当前节点时，该审批人<span>{ item.readBefore == 1 ? '能' : '不能' }</span>查看本表单</div>)
+                        } if (typeof item.readBefore != 'undefined') {
+                          temp.push(<div class="text">审批未到当前节点时，该审批人<span>{item.readBefore == 1 ? '能' : '不能'}</span>查看本表单</div>)
                         }
                         if ((Array.isArray(item.formAuthority) && item.formAuthority.length > 0) || (!Array.isArray(item.formAuthority) && item.formAuthority)) {
                           temp.push(<div class="text">表单权限：<span>{this.getFormRoleName(item.formAuthority, index)}</span></div>)
@@ -180,8 +241,8 @@ export default class extends tsx.Component<Vue> {
                         }
                         return (
                           <div>
-                            <ul class="message"> 
-                              {temp.map(current => <li><div class="pointer"></div>{current}</li>)} 
+                            <ul class="message">
+                              {temp.map(current => <li><div class="pointer"></div>{current}</li>)}
                             </ul>
                             {(index === 0 && temp.length !== 0) && <div class="notice">注：1、“已保存/已提交/审核中/已退回”的状态下不可删除。 <br />
   &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; 2、“已完成”的状态下删除，需要走审批流程。</div>}
@@ -192,7 +253,7 @@ export default class extends tsx.Component<Vue> {
                   </div>
                 </div>
                 {
-                  (index !== list.length - 1 || index == 0) && <div class="addNewBtn" onClick={this.addNewNode.bind(this, index)}>
+                  (index !== list.length - 1 || index == 0) && this.pageType !== pageTypeEnum.detail && <div class="addNewBtn" onClick={this.addNewNode.bind(this, index)}>
                     <img src={require('@/assets/img/addBtn.png')} />
                   </div>
                 }
@@ -204,12 +265,12 @@ export default class extends tsx.Component<Vue> {
         {
           this.showModel && (
             <customModel
-            formOptions={this.formOptions}
-            show={this.showModel}
-            title={this.modelTitle}
-            theme={this.theme}
-            on={{ ['update:show']: this.handleSync.bind(this, 'showModel'), submit: this.formSubmit }} >
-          </customModel>
+              formOptions={this.formOptions}
+              show={this.showModel}
+              title={this.modelTitle}
+              theme={this.theme}
+              on={{ ['update:show']: this.handleSync.bind(this, 'showModel'), submit: this.formSubmit }} >
+            </customModel>
           )
         }
       </div>
